% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLtable.R
\name{BLtable.layout}
\alias{BLtable.layout}
\title{Function to gather information about the BL-table before computing begins}
\usage{
BLtable.layout(data, info)
}
\arguments{
\item{data}{A data frame that contains the baseline (BL) data (one row per participant).}

\item{info}{A data frame with specifications for the BL-table. (See Examples below.)}
}
\value{
Data frame containing layout information of the table to be computed
   having these columns:
 \describe{
 \item{Variable, Table.label, Type, Test, etc.}{Inherited from data set submitted as 'info'}
 \item{n_rows}{Number of rows in the table for the corresponding variable}
 \item{first_row}{First table row used for the corresponding variable}
 \item{last_row}{Last table row used for the corresponding variable}
}
}
\description{
\code{BLtable.layout} is called by \code{\link{BLtable}} to gather information about the table before computing begins
}
\examples{
\dontrun{ Called by other function and not intended to be called by user directly. }
}
