% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{BMIgroup}
\alias{BMIgroup}
\title{Function to render BMI-groups based on BMI-values.}
\usage{
BMIgroup(BMI.values)
}
\arguments{
\item{BMI.values}{The input vector containing BMI-values.}
}
\value{
A vector of BMI groups as ordered factor.
}
\description{
\code{BMIgroup} takes a vector with BMI values and returns a vector with resulting BMI-groups (as ordered factor).
}
\examples{
BMIgroup(BMI.values = c(18.6, 22.5, 14.3, 28.4, 38.3))
}
