% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{CheckNames}
\alias{CheckNames}
\title{Function to test the identity of variable names in two data sets.}
\usage{
CheckNames(data1, data2, identical.order = TRUE,
  names.data1.only = FALSE)
}
\arguments{
\item{data1}{A data frame.}

\item{data2}{A data frame.}

\item{identical.order}{Whether the variables/columns are required to be in the same position
in both data sets.  The default is \code{identical.order = TRUE}, if \code{identical.order = FALSE}
the test is with respect to identical spelling only and irrespective of the relative position in the respective data frames.}

\item{names.data1.only}{Not yet implemented but intended as follows: If \code{names.data1.only = TRUE} testing is only with respect to the existence and spelling
of all variables in \code{data1} (useful for testing if a new version of a data set (\code{data2}) contains all of the
variables in an earlier version (\code{data1}) where it is understood that new variables may have been added to the new version).}
}
\value{
\code{TRUE} or \code{FALSE} depending on the testing result.
}
\description{
\code{CheckNames} takes two data frames and compares the included variable names.
}
\examples{
CheckNames(data1 = example_data, data2 = example_data) # Trivially \\code{TRUE}, of course.
}
