% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLtable.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Synthetic data for compiling a table with 2 groups}
\format{A data frame with 100 rows and 5 variables:
\describe{
 \item{ID}{Fictitious ID}
 \item{Group}{group membership - GroupA or GroupB (as text)}
 \item{Gender}{Fictitious gender as example of a binary variable (as text)}
 \item{Age}{Fictitious age in years as example of a continuous variable}
 \item{Some.measure}{Fictitious continuous measure with skewed distribution.}
 \item{Educ}{Fictitious education level as example of a factor (with 8 possibly unused factor levels)}
}}
\source{
Synthetic data created by the package author
}
\usage{
example_data
}
\description{
A dataset containing typical entries for baseline tables
}
\keyword{datasets}
