% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORtableGLMER.R
\docType{data}
\name{example_data2}
\alias{example_data2}
\title{Synthetic data for compiling a OR table}
\format{A data frame with 100 rows and 5 variables:
\describe{
 \item{ID}{Fictitious ID}
 \item{gender}{Fictitious gender as example of a binary variable (as text)}
 \item{trait}{Fictitious factor with three levels}
 \item{cov}{Fictitious continuous covariate}
 \item{outc}{Fictitious binary outcome (0/1)}
}}
\source{
Synthetic data created by the package author
}
\usage{
example_data2
}
\description{
A dataset containing typical entries for ORtable with IDs indicating grouping.
}
\keyword{datasets}
