% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLtable.R
\docType{data}
\name{example_variables}
\alias{example_variables}
\title{Description for compilation of table (participant data set)}
\format{A data frame with 3 rows and 5 columns:
\describe{
 \item{Variable}{Variable name}
 \item{BLtable}{Indicator whether to include in BL table}
 \item{Table.label}{Label to use in the created table}
 \item{Type}{What type of variable this is: Continuous, Dichotomous, or Factor?}
 \item{Test}{Test to be used for p-value: t-test, Wilcoxon, Chi-square, or Fisher?)}
}}
\source{
Synthetic data created by the package author
}
\usage{
example_variables
}
\description{
A dataset containing the variables and tests to be included in the baseline table
}
\keyword{datasets}
