% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORtableGLMER.R
\docType{data}
\name{example_variables2}
\alias{example_variables2}
\title{Description for compilation of table (participant data set)}
\format{A data frame with 3 rows and 5 columns:
\describe{
 \item{Variable}{Variable name}
 \item{ORtable}{Indicator whether to include in BL table}
 \item{Table.label}{Label to use in the created table}
 \item{Type}{What type of variable this is: Continuous, Dichotomous, or Factor?}
}}
\source{
Synthetic data created by the package author
}
\usage{
example_variables2
}
\description{
A dataset containing the variables and tests to be included in the baseline table
}
\keyword{datasets}
