% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLtable.R
\name{generateinputtable}
\alias{generateinputtable}
\title{Function to create a skeleton version of input table.}
\usage{
generateinputtable(inputset)
}
\arguments{
\item{inputset}{A data frame that contains the baseline (BL) data (one row per participant).
Intended use: same data set as used by \code{\link{BLtable}}.}
}
\value{
A data frame useful to create table-input for for \code{\link{BLtable}}.
}
\description{
\code{generateinputtable} takes an R data frame and creates table-input useful for \code{\link{BLtable}}.
}
\examples{
generateinputtable(example_data)
}
