% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLtable.R
\name{grouptest}
\alias{grouptest}
\title{Function to test the dichotomous variable identifying the two groups in the BL-table.}
\usage{
grouptest(data, group.var)
}
\arguments{
\item{data}{A data frame that contains the baseline (BL) data (one row per participant).}

\item{group.var}{The name of the variable (column) in \code{data} that partitions
the data set into two groups.}
}
\value{
Prints warning message and/or stops processing with error message printed.
}
\description{
Function to test the dichotomous variable identifying the two groups in the BL-table.
}
\examples{
Called by other function - not intended to be called by user directly.
}
