% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRtable.R
\name{HRtable.text}
\alias{HRtable.text}
\title{Function to compute and render text entries for HR-table}
\usage{
HRtable.text(this.data, this.outcome, this.var, this.type, sign.digits,
  sign.digits.HR, pvalue.digits, pvalue.cutoff, less.than.character)
}
\arguments{
\item{this.data}{Data used for the computations}

\item{this.outcome}{A text string as input for survival::coxph, e.g.,
"Surv(Follow_up_Years, Event_Death)" (incl. the quotes). This string will
be passed to coxph as the left-side of the model equation with
\code{as.formula(paste(this.outcome, " ~ ", this.var, sep = ""))}}

\item{this.var}{Variable in the table (must exist in \code{this.data})}

\item{this.type}{Variable type for table}

\item{sign.digits}{Digits used for parameter estimates}

\item{sign.digits.HR}{Digits used for hazard ratios and conf. intervals}

\item{pvalue.digits}{Digits used for p-values}

\item{pvalue.cutoff}{Cutoff for changing p-values to simply 'smaller than', e.g., p < 0.001}

\item{less.than.character}{The character to be prepended to the cutoff-value for the p-values (e.g., '< 0.001'). The default
is \code{less.than.character = "< "} but it can be reset to any character string. If the result is
passed on to html-rendering this symbol can be produced by \code{less.than.character = "&lt; "}
(HTML-friendly name for 'less than' since '<' is used in HTML-tags and might not render correctly as a character; e.g.,
when using function htmlTable() in the htmlTable-package).}
}
\value{
Output used by \code{HRtable()}
}
\description{
Function to compute and render text entries for HR-table
}
\examples{
Called by other function - not intended to be called by user directly.
}
