% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{ListVariables}
\alias{ListVariables}
\title{Function to summarize data in a data frame.}
\usage{
ListVariables(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A character matrix useful to print directly or as input to R-Markdown (e.g., using the kable-function) listing
 \describe{
 \item{Variable}{Variable name}
 \item{Missing entries (count)}{The number of (formally) missing values (note: 'coded missingness' such a -99 would not be counted as missing
       and would be an acceptable input value).}
 \item{Unique entries (count incl missing if applicable)}{The number of distinct entry-values, incl. formally missing as one such entry, if applicable.}
 \item{List of actual entries}{Listing actually occuring distinct values in order (typicaly alphabetic order).
       If there are 11 or more distinct entries the first and last three are printed separated by '(...)' to indicate additional entries in between.}
}
}
\description{
\code{ListVariables} takes a data frame and lists all existing variables incl. summary descriptions.
}
\examples{
ListVariables(data = example_data)
}
