% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONEtable.R
\name{ONEtable.layout}
\alias{ONEtable.layout}
\title{Function to gather information about the ONE-table before computing begins}
\usage{
ONEtable.layout(data, info)
}
\arguments{
\item{data}{A data frame that contains the data (one row per participant).}

\item{info}{A data frame with specifications for the ONE-table. (See Examples below.)}
}
\value{
Data frame containing layout information of the table to be computed
   having these columns:
 \describe{
 \item{Variable, Table.label, Type, etc.}{Inherited from data set submitted as 'info'}
 \item{n_rows}{Number of rows in the table for the corresponding variable}
 \item{first_row}{First table row used for the corresponding variable}
 \item{last_row}{Last table row used for the corresponding variable}
}
}
\description{
\code{ONEtable.layout} is called by \code{\link{ONEtable}} to gather information about the table before computing begins
}
\examples{
Called by other function - not intended to be called by user directly.
}
