% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONEtable.R
\name{ONEtable}
\alias{ONEtable}
\title{Function to compile ONE-table.}
\usage{
ONEtable(data, info, sign.digits = 2, sign.digits.prop = 1,
  show.missing.info = TRUE, test.input = FALSE,
  factor.level.bullet = "- ", linebreak.tag = "")
}
\arguments{
\item{data}{A data frame that contains the data (one row per participant).}

\item{info}{A data frame with specifications for the ONE-table. (See Examples below.)}

\item{sign.digits}{Digits used for mean, standard deviation, etc.}

\item{sign.digits.prop}{Digits used for proportion of cases for factors (and missing entries, if requested).}

\item{show.missing.info}{If \code{show.missing.info = TRUE} information about
missing information in a variable is added to the table. \code{show.missing.info = FALSE}
suppresses that information.}

\item{test.input}{Whether the input is tested for logical consistency.  Warning: testing is not
comprehensive.  Setting \code{test.input = FALSE} will bypass logical tests.}

\item{factor.level.bullet}{The character to be prepended to factor levels (listing). The default
is \code{factor.level.bullet = "- "} (hyphen/minus) but it can be reset to any character string. If the result is
passed on to html-rendering an en-dash can be produced by \code{factor.level.bullet = "&#8211; "} (HTML-code for the en-dash).
Other useful options for (subsequent) HTML-rendering: \code{factor.level.bullet = "&nbsp; &nbsp; "} forces non-breaking space and effectively
leads to indentation of factor levels; \code{factor.level.bullet = "&#187; "} prints right double angle quotes in HTML.}

\item{linebreak.tag}{Can be used to insert a linebreak-command in the table headers to bring the counts to a new line.
The default is \code{linebreak.tag = ""} effectively not inserting any such command.  For subsequent HTML-rendering
\code{linebreak.tag = "<br>"} (HTML tag to insert a single line break) is useful.}
}
\value{
An R character-matrix containing the compiled table information.
}
\description{
Function to compile ONE-table.
}
\examples{
ONEtable(data = example_data, info = example_variables, show.missing.info = TRUE)
}
