% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONEtable.R
\name{ONEtable.text}
\alias{ONEtable.text}
\title{Function to compute and render text entries for ONE-table}
\usage{
ONEtable.text(data, this.var, this.type, sign.digits, sign.digits.prop)
}
\arguments{
\item{data}{Data used for the computations}

\item{this.var}{Variable in the table (must exist in \code{data})}

\item{this.type}{Variable type for table}

\item{sign.digits}{Digits used for mean, standard deviation, etc.}

\item{sign.digits.prop}{Digits used for proportion of cases for factors (and missing entries, if requested).}
}
\value{
Output used by \code{ONEtable()}
}
\description{
Function to compute and render text entries for ONE-table
}
\examples{
Called by other function - not intended to be called by user directly.
}
