% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORtable.R
\name{ORtable.text}
\alias{ORtable.text}
\title{Function to compute and render text entries for OR-table}
\usage{
ORtable.text(this.data, this.outcome, this.var, this.type, sign.digits,
  sign.digits.OR, pvalue.digits, pvalue.cutoff, less.than.character)
}
\arguments{
\item{this.data}{Data used for the computations}

\item{this.outcome}{The name of the variable (column) in \code{data} that is
the binary outcome for the logistic regression models.}

\item{this.var}{Variable in the table (must exist in \code{this.data})}

\item{this.type}{Variable type for table}

\item{sign.digits}{Digits used for parameter estimate}

\item{sign.digits.OR}{Digits used for odds ratio with conf.intervals}

\item{pvalue.digits}{Digits used for p-values.}

\item{pvalue.cutoff}{Cutoff for changing p-values to simply 'smaller than', e.g., p < 0.001.}

\item{less.than.character}{The character to be prepended to the cutoff-value for the p-values (e.g., '< 0.001'). The default
is \code{less.than.character = "< "} but it can be reset to any character string. If the result is
passed on to html-rendering this symbol can be produced by \code{less.than.character = "&lt; "}
(HTML-friendly name for 'less than' since '<' is used in HTML-tags and might not render correctly as a character; e.g.,
when using function htmlTable() in the htmlTable-package).}
}
\value{
Output used by \code{ORtable()}
}
\description{
Function to compute and render text entries for OR-table
}
\examples{
Called by other function - not intended to be called by user directly.
}
