% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORtableGLMER.R
\name{ORtableGLMER.layout}
\alias{ORtableGLMER.layout}
\title{Function to gather information about the OR-table before computing begins}
\usage{
ORtableGLMER.layout(data, info)
}
\arguments{
\item{data}{A data frame that contains the data (one row per participant).}

\item{info}{A data frame with specifications for the OR-table. (See Examples below.)}
}
\value{
Data frame containing layout information of the table to be computed
   having these columns:
 \describe{
 \item{Variable, Table.label, etc.}{Inherited from data set submitted as 'info'}
 \item{n_rows}{Number of rows in the table for the corresponding variable}
 \item{first_row}{First table row used for the corresponding variable}
 \item{last_row}{Last table row used for the corresponding variable}
}
}
\description{
\code{ORtableGLMER.layout} is called by \code{\link{ORtableGLMER}} to gather information about the table before computing begins
}
\examples{
\dontrun{ Called by other function - not intended to be called by user directly. }
}
