% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORtableGLMER.R
\name{ORtableGLMER}
\alias{ORtableGLMER}
\title{Function to compile OR-table based on GLMER with provided IDs for grouping.}
\usage{
ORtableGLMER(data, this.outcome, this.ID, info, sign.digits = 2,
  sign.digits.OR = 3, pvalue.digits = 3, pvalue.cutoff = 0.001,
  test.input = FALSE, factor.level.bullet = "- ",
  less.than.character = "< ", linebreak.tag = "")
}
\arguments{
\item{data}{A data frame that contains the data (one row per participant).}

\item{this.outcome}{The name of the variable (column) in \code{data} that is
the binary outcome for the logistic regression models.}

\item{this.ID}{The name of the variable (column) in \code{data} that
identifies the same individual/speciman. The intended use is to handle
non-independence of observations from participants with repeated outcomes.
This variable is used in the called lme::glmer function as part of the
formula to specify a random effect.}

\item{info}{A data frame with specifications for the OR-table. (See Examples below.)}

\item{sign.digits}{Digits used for parameter estimate}

\item{sign.digits.OR}{Digits used for odds ratio with conf.intervals}

\item{pvalue.digits}{Digits used for p-values.}

\item{pvalue.cutoff}{Cutoff for changing p-values to simply 'smaller than', e.g., p < 0.001.}

\item{test.input}{Whether the input is tested for logical consistency.  Warning: testing is not
comprehensive.  Setting \code{test.input = FALSE} will bypass logical tests.}

\item{factor.level.bullet}{The character to be prepended to factor levels (listing). The default
is \code{factor.level.bullet = "- "} (hyphen/minus) but it can be reset to any character string. If the result is
passed on to html-rendering an en-dash can be produced by \code{factor.level.bullet = "&#8211; "} (HTML-code for the en-dash).
Other useful options for (subsequent) HTML-rendering: \code{factor.level.bullet = "&nbsp; &nbsp; "} forces non-breaking space and effectively
leads to indentation of factor levels; \code{factor.level.bullet = "&#187; "} prints right double angle quotes in HTML.}

\item{less.than.character}{The character to be prepended to the cutoff-value for the p-values (e.g., '< 0.001'). The default
is \code{less.than.character = "< "} but it can be reset to any character string. If the result is
passed on to html-rendering this symbol can be produced by \code{less.than.character = "&lt; "}
(HTML-friendly name for 'less than' since '<' is used in HTML-tags and might not render correctly as a character; e.g.,
when using function htmlTable() in the htmlTable-package).}

\item{linebreak.tag}{Can be used to insert a linebreak-command in the table headers to bring the counts to a new line.
The default is \code{linebreak.tag = ""} effectively not inserting any such command.  For subsequent HTML-rendering
\code{linebreak.tag = "<br>"} (HTML tag to insert a single line break) is useful.}
}
\value{
An R character-matrix containing the compiled table information.
}
\description{
Function to compile OR-table based on GLMER with provided IDs for grouping.
}
\examples{
   ORtableGLMER(data=example_data2, this.outcome="outc", this.ID="ID",
     info=example_variables2,
     sign.digits=2, sign.digits.OR=3, pvalue.digits=3, pvalue.cutoff=0.001,
     test.input = FALSE, factor.level.bullet = "- ",
     less.than.character = "< ", linebreak.tag = "")
}
