% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLtable.R
\name{table.text}
\alias{table.text}
\title{Function to compute and render text entries for BL-table}
\usage{
table.text(this.data, this.var, group.var, this.GroupA, this.GroupB,
  this.type, this.test, sign.digits, sign.digits.prop, pvalue.digits,
  pvalue.cutoff, less.than.character)
}
\arguments{
\item{this.data}{Data used for the computations}

\item{this.var}{Variable in the table (must exist in \code{this.data})}

\item{group.var}{The name of the variable (column) in \code{data} that partitions
the data set into two groups.}

\item{this.GroupA}{Name (factor level) of 'Group A' in the table}

\item{this.GroupB}{Name (factor level) of 'Group B' in the table}

\item{this.type}{Variable type for table}

\item{this.test}{Test that determines p-value}

\item{sign.digits}{Digits used for mean, standard deviation, etc.}

\item{sign.digits.prop}{Digits used for proportion of cases for factors.}

\item{pvalue.digits}{Digits used for p-values.}

\item{pvalue.cutoff}{Cutoff for changing p-values to simply 'smaller than', e.g., p < 0.001.}

\item{less.than.character}{The character to be prepended to the cutoff-value for the p-values (e.g., '< 0.001'). The default
is \code{less.than.character = "< "} but it can be reset to any character string. If the result is
passed on to html-rendering this symbol can be produced by \code{less.than.character = "&lt; "}
(HTML-friendly name for 'less than' since '<' is used in HTML-tags and might not render correctly as a character; e.g.,
when using function htmlTable() in the htmlTable-package).}
}
\value{
Output used by \code{BLtable()}
}
\description{
Function to compute and render text entries for BL-table
}
\examples{
\dontrun{ Called by other function - not intended to be called by user directly. }
}
